refactor(loon-cli): split command runner by command family
